Update streamlit.app.py
put the list of selected fruits into a variable called fruits_selected. 
use the fruits in our fruits_selected list to pull rows from the full data set (and assign that data to a variable called fruits_to_show). 
use the data in fruits_to_show in the dataframe it displays on the page.